requirements.txt
streamlit
easyocr
markitdown
pdf2image
pillow
numpy